# swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is a sample server Petstore server.  You can find out more about    
    Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,
    #swagger](http://swagger.io/irc/).      For this sample, you can use the api
    key `special-key` to test the authorization     filters.
  version: 1.0.0
  title: Swagger Petstore
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: www.yangmz.com
basePath: /
...
schemes:
  - http
paths:
...
  '/user/{userId}':
    get:
      tags:
        - user
      summary: 获取用户信息
      description: ''
      operationId: getUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          type: number
      responses:
        '405':
          description: Invalid input
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /app/uploadArticle/image:
    post:
      tags:
        - app article
      summary: 上传图片
      description: ''
      operationId: uploadImage
      consumes:
        - application/json
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: userId
          in: query
          description: ID of user
          required: true
          type: string
        - name: token
          in: query
          description: token or user
          required: true
          type: string
        - name: index
          in: formData
          type: string
          description: The file index.
        - name: file
          in: formData
          type: file
          description: The file to upload.
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/UserToken'
        '405':
          description: Invalid input
...
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  UserToken:
    type: object
    properties:
      id:
        type: integer
        format: int64
      userName:
        type: string
      userPicAddress:
        type: string
      token:
        type: string
    xml:
      name: UserToken
...